restructure dev environment
