+++ .gitignore
@@ -1,5 +1,2 @@
 # Ignore res.json since this has ID and Token info for Oanda Account
 res.json
-
-# Ignore password for connecting to a local instance of PostgreSQL
-connect/password.go
