+++ cmd/go.mod
@@ -1,0 +1,13 @@
+module github.com/davidhintelmann/Oanda-Go/cmd
+
+go 1.22.5
+
+require github.com/davidhintelmann/Oanda-Go/restful v0.0.0-20230516023247-0b3e7cc616a1
+
+require (
+	github.com/golang-sql/civil v0.0.0-20190719163853-cb61b32ac6fe // indirect
+	github.com/golang-sql/sqlexp v0.1.0 // indirect
+	github.com/microsoft/go-mssqldb v0.21.0 // indirect
+	golang.org/x/crypto v0.6.0 // indirect
+	gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce // indirect
+)
