--- go.mod
@@ -1,23 +1,0 @@
-module github.com/davidhintelmann/Oanda-Go
-
-go 1.20
-
-require (
-	github.com/davidhintelmann/Oanda-Go/connect v0.0.0-20230513072756-f6694b5ab948
-	github.com/davidhintelmann/Oanda-Go/restful v0.0.0-20230513072756-f6694b5ab948
-	github.com/jackc/pgx/v5 v5.6.0
-)
-
-require (
-	github.com/golang-sql/civil v0.0.0-20190719163853-cb61b32ac6fe // indirect
-	github.com/golang-sql/sqlexp v0.1.0 // indirect
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jackc/puddle/v2 v2.2.1 // indirect
-	github.com/microsoft/go-mssqldb v0.21.0 // indirect
-	github.com/stretchr/testify v1.9.0 // indirect
-	golang.org/x/crypto v0.24.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
-	golang.org/x/text v0.16.0 // indirect
-	gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce // indirect
-)
