+++ restful/go.mod
@@ -1,12 +1,3 @@
 module github.com/davidhintelmann/Oanda-Go/restful
 
 go 1.20
-
-require github.com/microsoft/go-mssqldb v0.21.0
-
-require (
-	github.com/golang-sql/civil v0.0.0-20190719163853-cb61b32ac6fe // indirect
-	github.com/golang-sql/sqlexp v0.1.0 // indirect
-	golang.org/x/crypto v0.24.0 // indirect
-	gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce // indirect
-)
